Add dju computation

+++ api/models/Dju.js
@@ -3,7 +3,7 @@ module.exports = {
         day: {type: 'integer', required: true},
         month: {type: 'integer', required: true},
         hour: {type: 'integer', required: true},
-        temperature: {type: 'integer', required: true},
+        temperature: {type: 'float', required: true},
         label: 'string'
     }
 };
